# perf: rewrite CoreTable.complete to be more performant