--- a/package.json
+++ b/package.json
@@ -128,7 +128,6 @@
         "express": "^4.18.1",
         "express-async-errors": "^3.1.1",
         "express-rate-limit": "^5.1.3",
-        "fast-cartesian": "^5.1.0",
         "filenamify": "^4.1.0",
         "fortune": "^5.5.17",
         "fs-extra": "^11.1.1",