--- a/packages/@ourworldindata/core-table/package.json
+++ b/packages/@ourworldindata/core-table/package.json
@@ -16,8 +16,7 @@
   "dependencies": {
     "@ourworldindata/utils": "workspace:^",
     "d3": "^6.1.1",
-    "dayjs": "^1.11.5",
-    "fast-cartesian": "^5.1.0"
+    "dayjs": "^1.11.5"
   },
   "devDependencies": {
     "@types/d3": "^6",