--- a/packages/@ourworldindata/core-table/src/CoreTable.ts
+++ b/packages/@ourworldindata/core-table/src/CoreTable.ts
@@ -57,7 +57,6 @@ import {
     getDropIndexes,
     parseDelimited,
     rowsFromMatrix,
-    cartesianProduct,
     sortColumnStore,
     emptyColumnsInFirstRowInDelimited,
     truncate,