--- a/packages/@ourworldindata/core-table/src/index.ts
+++ b/packages/@ourworldindata/core-table/src/index.ts
@@ -107,7 +107,6 @@ export {
     isCellEmpty,
     trimEmptyRows,
     trimArray,
-    cartesianProduct,
     sortColumnStore,
     emptyColumnsInFirstRowInDelimited,
 } from "./CoreTableUtils.js"