--- a/packages/@ourworldindata/core-table/src/index.ts
+++ b/packages/@ourworldindata/core-table/src/index.ts
@@ -88,7 +88,6 @@ export {
     toleranceInterpolation,
     interpolateRowValuesWithTolerance,
     makeKeyFn,
-    appendRowsToColumnStore,
     concatColumnStores,
     rowsToColumnStore,
     autodetectColumnDefs,