While working on Cloudflare Workers, I realized that `CoreTable.complete` takes up a non-trivial amount of time.

Seeing that we only ever call `complete` with exactly 2 columns, and that the current implementation was written in a way that supports an arbitrary number of cols (which makes everything trickier), I rewrote it to only ever support 2 columns, but in a way more performant way.

I benchmarked this on https://ourworldindata.org/grapher/life-expectancy-vs-gdp-per-capita:
The two invocations of `complete()` took **385ms** before; and now take **85ms**. And that's on a very fast M2 Max; I expect the difference to be similarly large on other machines too.

Here's also a (manual) timing output for localhost:
```
http://localhost:3030/grapher/life-expectancy-vs-gdp-per-capita

15:56:52.155 CoreTable.ts:1452 c: 0.009765625 ms got cols
15:56:52.168 CoreTable.ts:1474 c: 12.537109375 ms filled map
15:56:52.174 CoreTable.ts:1482 c: 18.829833984375 ms removed rows
15:56:52.184 CoreTable.ts:1494 c: 28.27490234375 ms built rows
15:56:52.219 CoreTable.ts:1499 c: 63.19384765625 ms appended rows
15:56:52.219 CoreTable.ts:1500 c: 63.3369140625 ms
---
# Second call, this time on an already-complete table
15:56:52.285 CoreTable.ts:1452 c: 0.005126953125 ms got cols
15:56:52.306 CoreTable.ts:1460 c: 21.652099609375 ms skip 256634 256634
15:56:52.306 CoreTable.ts:1461 c: 21.774169921875 ms
```

Another big performance difference can be seen in the SVG Tester; on current master the "Generating SVGs" phase is taking **17-18mins**, on this branch it's taking around **10-13mins**. ⚡ 

## TODOs
- [x] Implement daniel's suggestion below
- [x] Check: Why are we often running `complete()` twice on the same table? Where the second one is already complete?
  - => We're running up to three `interpolate...()` calls in `ScatterPlot.transformTable()`
- [x] `appendRows()` is still taking a long time, can we optimize it further?