@marcelgerber 2023-09-06
That the SVG Tester is now showing any differences is very promising already!