@marcelgerber 2023-09-06
Also tested this with one of the worst offenders now: https://ourworldindata.org/grapher/GDP-vs-manufacturing-employment

On live it's **587ms**, with this PR it's **151ms**.