@danyx23 2023-09-06
Very nice! I thought about this problem a bit over the last few days and I think it would make sense to trim "auxilliary" dimensions to the time range of the "main" dimensions - that should drastically reduce the number of elements in the full cartesian product for scatters with long run population used as size. I'll try to do that independently of your implementation here but that should give us an additional small speed boost for scatters.