@marcelgerber 2023-09-06
Another thing to check: Why are we often running `complete` twice on the same (?) table? Where one of the tables is already complete, and we now return early?